+++ OrderService/Models/OrderDish.cs
@@ -4,6 +4,6 @@ namespace OrderService.Models{
         public int OrderId { get; set; }
         public int DishId { get; set; }
         public int Quantity { get; set; }
-        public int CurrentPrice { get; set; }
+        public decimal CurrentPrice { get; set; }
     }
 }
